+++ XFaction/Core/Modules/Confederate/Guild.lua
@@ -11,6 +11,8 @@ function Guild:new()
     self.__name = ObjectName
 
     self._Key = nil
+    self._ID = nil
+    self._StreamID = nil
     self._Name = nil
     self._Initials = nil
     self._Faction = nil
@@ -40,6 +42,8 @@ function Guild:Print()
     XFG:DoubleLine(LogCategory)
     XFG:Debug(LogCategory, ObjectName .. " Object")
     XFG:Debug(LogCategory, "  _Key (" .. type(self._Key) .. "): ".. tostring(self._Key))
+    XFG:Debug(LogCategory, "  _ID (" .. type(self._ID) .. "): ".. tostring(self._ID))
+    XFG:Debug(LogCategory, "  _StreamID (" .. type(self._StreamID) .. "): ".. tostring(self._StreamID))
     XFG:Debug(LogCategory, "  _Name (" .. type(self._Name) .. "): ".. tostring(self._Name))
     XFG:Debug(LogCategory, "  _Initials (" .. type(self._Initials) .. "): ".. tostring(self._Initials))
     XFG:Debug(LogCategory, "  _Faction (" .. type(self._Faction) .. ")")
@@ -78,8 +82,36 @@ function Guild:SetInitials(inInitials)
     return self:GetInitials()
 end
 
+function Guild:HasID()
+    return self._ID ~= nil
+end
+
+function Guild:GetID()
+    return self._ID
+end
+
+function Guild:SetID(inID)
+    assert(type(inID) == 'number')
+    self._ID = inID
+    return self:GetID()
+end
+
 function Guild:GetFaction()
     return self._Faction
+end
+
+function Guild:HasStreamID()
+    return self._StreamID ~= nil
+end
+
+function Guild:GetStreamID()
+    return self._StreamID
+end
+
+function Guild:SetStreamID(inStreamID)
+    assert(type(inStreamID) == 'number')
+    self._StreamID = inStreamID
+    return self:GetStreamID()
 end
 
 function Guild:SetFaction(inFaction)